[ISSUE #5089] Delegate plugin store configuration by broker
